+++ e3nn_core/__init__.py
@@ -4,8 +4,8 @@ from .cg import ClebschGordanKey, clebsch_gordan, su2_clebsch_gordan, wigner_3j
 from .instructions import TensorProductInstruction, generate_tensor_product_instructions, make_tensor_product_instructions
 from .irreps import Irrep, Irreps, MulIrrep, SortResult
 from .normalization import NormalizationMetadata
-from .rotations import RotationAngles
-from .wigner import WignerDKey, Wigner3jKey, change_basis_real_to_complex, so3_generators, su2_generators
+from .runtime import register_runtime, registered_runtimes
+from .wigner import change_basis_real_to_complex, so3_generators, su2_generators
 
 __all__ = [
     "ClebschGordanKey",
@@ -13,11 +13,10 @@ __all__ = [
     "Irreps",
     "MulIrrep",
     "NormalizationMetadata",
-    "RotationAngles",
+    "register_runtime",
+    "registered_runtimes",
     "SortResult",
     "TensorProductInstruction",
-    "Wigner3jKey",
-    "WignerDKey",
     "change_basis_real_to_complex",
     "clebsch_gordan",
     "generate_tensor_product_instructions",